+++ kuka_sunrise_control/robot_control/package.xml
@@ -13,6 +13,7 @@
   <build_depend>sensor_msgs</build_depend>
   <build_depend>rclcpp_lifecycle</build_depend>
   <build_depend>kuka_sunrise_interfaces</build_depend>
+  <build_depend>kroshu_ros2_core</build_depend>
   
   <exec_depend>rclcpp</exec_depend>
   <exec_depend>sensor_msgs</exec_depend>
